requirements.txt
Flask==3.0.0
flask-cors==4.0.0
flask-socketio==5.3.5
flask-sqlalchemy==3.1.1
flask-mail==0.10.0
flask-limiter==3.5.0
psycopg2-binary==2.9.11
python-dotenv==1.0.0
cryptography==42.0.0
redis==5.0.1
google-generativeai==0.3.2
openai>=1.40.0
python-socketio==5.10.0
bleach==6.1.0
email-validator==2.1.0
